[重構] 拆分 postOrder 寄信區塊，放到 config/mailer.js
